Relaxed and simplified the route acquisition algorithm.
The only requirement is that an unfeasible route should never be selected;
thus, it is okay to create an entry for an unfeasible update, it is just
illegal to select it.  This also changes the behaviour for a router-id
change: instead of treating it as a retraction, we treat it as a normal
update, but send an update straight away.

Also added a note about keeping unfeasible routes in the routing table,
since that confused at least one implementer.